# internal/handler/user_handler.go
package handler

import (
	"net/http"
	"strconv"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/golang-jwt/jwt/v5"
	"github.com/sandi/lumiina/internal/model"
	"github.com/sandi/lumiina/internal/service"
)

type UserHandler struct {
	service   service.UserService
	jwtSecret string
}

func NewUserHandler(service service.UserService, jwtSecret string) *UserHandler {
	return &UserHandler{service: service, jwtSecret: jwtSecret}
}

// Register handles new user registration and dispatches email verification.
// @Summary Register a new user
// @Description Creates a new user account and returns a verification token.
// @Tags auth
// @Accept json
// @Produce json
// @Param req body model.RegisterRequest true "User Registration Data"
// @Success 201 {object} map[string]interface{}
// @Router /auth/register [post]
func (h *UserHandler) Register(c *gin.Context) {
	var req model.RegisterRequest

	err := c.ShouldBindJSON(&req)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	if req.Password != req.ConfirmPassword {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Konfirmasi password tidak cocok dengan password yang dimasukkan"})
		return
	}

	user := model.User{
		Username: req.Username,
		Email:    req.Email,
		Password: req.Password,
	}

	err = h.service.Register(&user)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusCreated, gin.H{
		"message": "Registrasi berhasil! Tautan verifikasi telah dikirimkan ke email Anda.",
		"data":    user,
	})
}

// Login authenticates a user and returns a signed JWT access token.
// @Summary Authenticate user
// @Description Logs in using email or username along with password, returning a JWT Bearer token.
// @Tags auth
// @Accept json
// @Produce json
// @Param req body model.LoginRequest true "User Credentials"
// @Success 200 {object} map[string]interface{} "Returns JWT Bearer token and user info"
// @Failure 400 {object} map[string]string "Invalid input payload"
// @Failure 401 {object} map[string]string "Invalid credentials or unverified account"
// @Router /auth/login [post]
func (h *UserHandler) Login(c *gin.Context) {
	var loginRequest model.LoginRequest

	err := c.ShouldBindJSON(&loginRequest)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": err.Error()})
		return
	}

	user, err := h.service.Login(loginRequest.Identifier, loginRequest.Password)
	if err != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": err.Error()})
		return
	}

	token := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
		"user_id": user.ID,
		"role":    user.Role,
		"exp":     time.Now().Add(time.Hour * 24).Unix(),
	})

	secret := h.jwtSecret

	tokenString, err := token.SignedString([]byte(secret))
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to generate token"})
		return
	}

	c.JSON(http.StatusOK, gin.H{
		"message": "Login successful",
		"token":   tokenString,
	})
}

// VerifyEmail validates the email confirmation token and activates the account.
// @Summary Verify email address
// @Description Activates the user account using the single-use token sent via email.
// @Tags auth
// @Produce json,html
// @Param token query string true "Verification Token"
// @Success 200 {object} map[string]string "Account successfully activated"
// @Failure 400 {object} map[string]string "Missing, expired, or invalid token"
// @Router /auth/verify-email [get]
func (h *UserHandler) VerifyEmail(c *gin.Context) {
	token := c.Query("token")
	wantsJSON := c.GetHeader("Accept") == "application/json"

	if token == "" {
		if wantsJSON {
			c.JSON(http.StatusBadRequest, gin.H{"error": "Token verifikasi wajib disertakan"})
			return
		}
		c.Data(http.StatusBadRequest, "text/html; charset=utf-8", []byte(renderVerificationErrorPage("Token Tidak Ditemukan", "Tautan verifikasi tidak lengkap atau tidak memiliki parameter token yang valid.")))
		return
	}

	err := h.service.VerifyEmail(token)
	if err != nil {
		if wantsJSON {
			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
			return
		}
		c.Data(http.StatusBadRequest, "text/html; charset=utf-8", []byte(renderVerificationErrorPage("Tautan Kedaluwarsa / Tidak Valid", err.Error())))
		return
	}

	if wantsJSON {
		c.JSON(http.StatusOK, gin.H{
			"message": "Email berhasil diverifikasi! Akun Anda kini sudah aktif. Silakan login.",
		})
		return
	}

	c.Data(http.StatusOK, "text/html; charset=utf-8", []byte(renderVerificationSuccessPage()))
}

// ForgotPassword initiates the password reset workflow.
// @Summary Request password reset
// @Description Sends a 15-minute ephemeral password reset token to the given email address.
// @Tags auth
// @Accept json
// @Produce json
// @Param req body model.ForgotPasswordRequest true "Registered Email Address"
// @Success 200 {object} map[string]string "Generic success response to prevent account enumeration"
// @Failure 400 {object} map[string]string "Invalid email format"
// @Router /auth/forgot-password [post]
func (h *UserHandler) ForgotPassword(c *gin.Context) {
	var req model.ForgotPasswordRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	err := h.service.ForgotPassword(req.Email)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Gagal memproses permintaan reset password"})
		return
	}

	c.JSON(http.StatusOK, gin.H{
		"message": "Jika alamat email terdaftar di sistem kami, instruksi reset kata sandi telah dikirimkan ke email Anda.",
	})
}

// ResetPassword updates the account password using the reset token.
// @Summary Reset account password
// @Description Sets a new password using a valid reset token.
// @Tags auth
// @Accept json
// @Produce json
// @Param req body model.ResetPasswordRequest true "Token and New Password"
// @Success 200 {object} map[string]string "Password reset successfully"
// @Failure 400 {object} map[string]string "Invalid token or password mismatch"
// @Router /auth/reset-password [post]
func (h *UserHandler) ResetPassword(c *gin.Context) {
	var req model.ResetPasswordRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	if req.NewPassword != req.ConfirmPassword {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Konfirmasi password baru tidak cocok"})
		return
	}

	err := h.service.ResetPassword(req.Token, req.NewPassword)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, gin.H{
		"message": "Kata sandi berhasil diperbarui! Silakan login menggunakan kata sandi baru Anda.",
	})
}

// GetMe returns the profile of the currently authenticated user.
// @Summary Get current user profile
// @Description Fetches details of the authenticated user from the JWT token.
// @Tags users
// @Produce json
// @Security BearerAuth
// @Success 200 {object} map[string]interface{} "User profile details"
// @Failure 401 {object} map[string]string "Unauthorized"
// @Failure 404 {object} map[string]string "User not found"
// @Router /users/me [get]
func (h *UserHandler) GetMe(c *gin.Context) {
	userIDFloat, exists := c.Get("user_id")
	if !exists {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
		return
	}
	userID := uint(userIDFloat.(float64))

	user, err := h.service.GetProfileByID(userID)
	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "User profile not found"})
		return
	}

	c.JSON(http.StatusOK, gin.H{"data": user})
}

// SearchUsers searches for artists/users by username.
// @Summary Search users & artists
// @Description Search for community artists by username with pagination.
// @Tags users
// @Produce json
// @Param q query string false "Search keyword"
// @Param page query int false "Page number (default: 1)"
// @Param limit query int false "Items per page (default: 20)"
// @Success 200 {object} map[string]interface{} "Paginated list of users"
// @Failure 500 {object} map[string]string "Database error"
// @Router /users/search [get]
func (h *UserHandler) SearchUsers(c *gin.Context) {
	q := c.Query("q")
	limitStr := c.DefaultQuery("limit", "20")
	pageStr := c.DefaultQuery("page", "1")

	limit, _ := strconv.Atoi(limitStr)
	page, _ := strconv.Atoi(pageStr)
	offset := (page - 1) * limit

	users, total, err := h.service.SearchUsers(q, limit, offset)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to search users"})
		return
	}

	c.JSON(http.StatusOK, gin.H{
		"page":  page,
		"limit": limit,
		"total": total,
		"data":  users,
	})
}

// GetUserProfile returns the public profile and artworks of a user by ID.
// @Summary Get public user profile
// @Description Fetches public artist profile and their uploaded artworks.
// @Tags users
// @Produce json
// @Param id path int true "User ID"
// @Success 200 {object} map[string]interface{} "Public artist profile"
// @Failure 400 {object} map[string]string "Invalid user ID"
// @Failure 404 {object} map[string]string "User not found"
// @Router /users/{id} [get]
func (h *UserHandler) GetUserProfile(c *gin.Context) {
	idStr := c.Param("id")
	id, err := strconv.Atoi(idStr)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid user ID"})
		return
	}

	user, err := h.service.GetProfileByID(uint(id))
	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "User profile not found"})
		return
	}

	c.JSON(http.StatusOK, gin.H{"data": user})
}